Add Boost Multiprecision Compatibility

+++ test/Integration.cpp
@@ -3,12 +3,12 @@
 #include <cmath>
 
 TEST ( Integration, Polinomial ) {
-    Function f = [](real x)->real{ return std::exp((-1)*std::pow(x, 2)); };
+    Function f = [](real x)->real{ return exp((-1)*pow(x, 2)); };
 
-    EXPECT_NEAR( computeIntegral( f, 0, 1,  1, Polinomial ), 0.778801, 1e-6 );
-    EXPECT_NEAR( computeIntegral( f, 0, 1,  2, Polinomial ), 0.68394,  1e-5 );
-    EXPECT_NEAR( computeIntegral( f, 0, 1,  3, Polinomial ), 0.74718,  1e-5 );
-    EXPECT_NEAR( computeIntegral( f, 0, 1,  4, Polinomial ), 0.74699,  1e-5 );
-    EXPECT_NEAR( computeIntegral( f, 0, 1,  5, Polinomial ), 0.74683,  1e-5 );
-    EXPECT_NEAR( computeIntegral( f, 0, 1, 10, Polinomial ), 0.74682,  1e-5 );
+    EXPECT_NEAR( static_cast<double>(computeIntegral( f, 0, 1,  1, Polinomial )), 0.778801, 1e-6 );
+    EXPECT_NEAR( static_cast<double>(computeIntegral( f, 0, 1,  2, Polinomial )), 0.68394,  1e-5 );
+    EXPECT_NEAR( static_cast<double>(computeIntegral( f, 0, 1,  3, Polinomial )), 0.74718,  1e-5 );
+    EXPECT_NEAR( static_cast<double>(computeIntegral( f, 0, 1,  4, Polinomial )), 0.74699,  1e-5 );
+    EXPECT_NEAR( static_cast<double>(computeIntegral( f, 0, 1,  5, Polinomial )), 0.74683,  1e-5 );
+    EXPECT_NEAR( static_cast<double>(computeIntegral( f, 0, 1, 10, Polinomial )), 0.74682,  1e-5 );
 }
